+++ src/CreativeCommons/icons/CCIconsProps.ts
@@ -1,0 +1,3 @@
+export type CCIconsProps = {
+  size: string | number;
+};